Add new function to get the client IP address
This removes the need to pelt the Gitea server with multiple GET
requests. The IP address is only detected when the client laptop is
docked on the the sync station and will only request updates in that
state. As soon as the laptop is removed from the dock, the class server
will stop trying to send network requests. This also removes the deinit
function as it should no longer be needed after this update is ready to
go live.