FEAT: Add middleware limiting access outside the localhost
This middleware function prevents the access of the students' server
to anyone on the outside network. This prevents the possibility of data transfer
from system to system as prohibited by DOC. Note: This can only be controlled
via the source code. After compile, This cannot be circumvented from the executable itself.