FEAT: Add website link API changes to include orphaned sites
This commit adds checking against the local copy of the repo dump and
Gitea API. I have also added handling for sites that are not a part of
the websites not hosted on Gitea. The untracked websites have basic
metadata included. This should allow for user notification and the
deletion or serving of said websites.

Still needed to do: Implement notification of updates to websites hosted
on Gitea.