FEAT: Add update to websites metadata locally
Updates info.json and reflects the proper website state allowing the
frontend to display pertinent alerts to the user. The migration script is
also embedded into the application and will be copied to the user's temp
folder and then executed. The function to run the script can be found in
filesystem.go. The script will be available until the next minor release
where it will be deprecated and removed.